+++ Realm/Realm/Handles/ObjectHandle.cs
@@ -1,4 +1,4 @@
-﻿////////////////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////////
 //
 // Copyright 2016 Realm Inc.
 //
@@ -52,10 +52,6 @@ namespace Realms
             public static extern void set_string(ObjectHandle handle, IntPtr propertyIndex,
                 [MarshalAs(UnmanagedType.LPWStr)] string value, IntPtr valueLen, out NativeException ex);
 
-            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_set_string_unique", CallingConvention = CallingConvention.Cdecl)]
-            public static extern void set_string_unique(ObjectHandle handle, IntPtr propertyIndex,
-                [MarshalAs(UnmanagedType.LPWStr)] string value, IntPtr valueLen, out NativeException ex);
-
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_get_string", CallingConvention = CallingConvention.Cdecl)]
             public static extern IntPtr get_string(ObjectHandle handle, IntPtr propertyIndex,
                 IntPtr buffer, IntPtr bufsize, [MarshalAs(UnmanagedType.I1)] out bool isNull, out NativeException ex);
@@ -78,9 +74,6 @@ namespace Realms
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_set_null", CallingConvention = CallingConvention.Cdecl)]
             public static extern void set_null(ObjectHandle handle, IntPtr propertyIndex, out NativeException ex);
 
-            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_set_null_unique", CallingConvention = CallingConvention.Cdecl)]
-            public static extern void set_null_unique(ObjectHandle handle, IntPtr propertyIndex, out NativeException ex);
-
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_set_bool", CallingConvention = CallingConvention.Cdecl)]
             public static extern void set_bool(ObjectHandle handle, IntPtr propertyIndex, IntPtr value, out NativeException ex);
 
@@ -92,9 +85,6 @@ namespace Realms
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_set_int64", CallingConvention = CallingConvention.Cdecl)]
             public static extern void set_int64(ObjectHandle handle, IntPtr propertyIndex, Int64 value, out NativeException ex);
-
-            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_set_int64_unique", CallingConvention = CallingConvention.Cdecl)]
-            public static extern void set_int64_unique(ObjectHandle handle, IntPtr propertyIndex, Int64 value, out NativeException ex);
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "object_add_int64", CallingConvention = CallingConvention.Cdecl)]
             public static extern void add_int64(ObjectHandle handle, IntPtr propertyIndex, Int64 value, out NativeException ex);
@@ -269,8 +259,10 @@ namespace Realms
                 throw new ArgumentNullException(nameof(value), "Object identifiers cannot be null");
             }
 
-            NativeMethods.set_string_unique(this, propertyIndex, value, (IntPtr)value.Length, out var nativeException);
-            nativeException.ThrowIfNecessary();
+            if (GetString(propertyIndex) != value)
+            {
+                throw new InvalidOperationException("Once set, primary key properties may not be modified.");
+            }
         }
 
         public string GetString(IntPtr propertyIndex)
@@ -375,23 +367,18 @@ namespace Realms
 
         public void SetInt64Unique(IntPtr propertyIndex, long value)
         {
-            NativeMethods.set_int64_unique(this, propertyIndex, value, out var nativeException);
-            nativeException.ThrowIfNecessary();
+            if (GetInt64(propertyIndex) != value)
+            {
+                throw new InvalidOperationException("Once set, primary key properties may not be modified.");
+            }
         }
 
         public void SetNullableInt64Unique(IntPtr propertyIndex, long? value)
         {
-            NativeException nativeException;
-            if (value.HasValue)
+            if (GetNullableInt64(propertyIndex) != value)
             {
-                NativeMethods.set_int64_unique(this, propertyIndex, value.Value, out nativeException);
+                throw new InvalidOperationException("Once set, primary key properties may not be modified.");   
             }
-            else
-            {
-                NativeMethods.set_null_unique(this, propertyIndex, out nativeException);
-            }
-
-            nativeException.ThrowIfNecessary();
         }
 
         public long GetInt64(IntPtr propertyIndex)