+++ Realm/Realm/Handles/SharedRealmHandle.cs
@@ -1,4 +1,4 @@
-﻿////////////////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////////
 //
 // Copyright 2016 Realm Inc.
 //
@@ -89,6 +89,26 @@ namespace Realms
 
             [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_write_copy", CallingConvention = CallingConvention.Cdecl)]
             public static extern void write_copy(SharedRealmHandle sharedRealm, [MarshalAs(UnmanagedType.LPWStr)] string path, IntPtr path_len, byte[] encryptionKey, out NativeException ex);
+
+            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_create_object", CallingConvention = CallingConvention.Cdecl)]
+            public static extern IntPtr create_object(SharedRealmHandle sharedRealm, TableHandle table, out NativeException ex);
+
+            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_create_object_int_unique", CallingConvention = CallingConvention.Cdecl)]
+            public static extern IntPtr create_object_unique(SharedRealmHandle sharedRealm, TableHandle table, long key,
+                                                             [MarshalAs(UnmanagedType.I1)] bool is_nullable,
+                                                             [MarshalAs(UnmanagedType.I1)] bool update,
+                                                             [MarshalAs(UnmanagedType.I1)] out bool is_new, out NativeException ex);
+
+            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_create_object_string_unique", CallingConvention = CallingConvention.Cdecl)]
+            public static extern IntPtr create_object_unique(SharedRealmHandle sharedRealm, TableHandle table,
+                                                             [MarshalAs(UnmanagedType.LPWStr)] string value, IntPtr valueLen,
+                                                             [MarshalAs(UnmanagedType.I1)] bool update,
+                                                             [MarshalAs(UnmanagedType.I1)] out bool is_new, out NativeException ex);
+
+            [DllImport(InteropConfig.DLL_NAME, EntryPoint = "shared_realm_create_object_null_unique", CallingConvention = CallingConvention.Cdecl)]
+            public static extern IntPtr create_object_unique(SharedRealmHandle sharedRealm, TableHandle table,
+                                                             [MarshalAs(UnmanagedType.I1)] bool update,
+                                                             [MarshalAs(UnmanagedType.I1)] out bool is_new, out NativeException ex);
         }
 
         [Preserve]
@@ -223,6 +243,69 @@ namespace Realms
         {
             NativeMethods.write_copy(this, path, (IntPtr)path.Length, encryptionKey, out var nativeException);
             nativeException.ThrowIfNecessary();
+        }
+
+        public IntPtr CreateObject(TableHandle table)
+        {
+            NativeException ex;
+            var result = NativeMethods.create_object(this, table, out ex);
+            ex.ThrowIfNecessary();
+            return result;
+        }
+
+        public IntPtr CreateObjectWithPrimaryKey(Property pkProperty, object primaryKey, TableHandle table, string parentType, bool update, out bool isNew)
+        {
+            switch (pkProperty.Type.UnderlyingType())
+            {
+                case PropertyType.String:
+                    if (primaryKey == null)
+                    {
+                        throw new ArgumentNullException(nameof(primaryKey), "Object identifiers cannot be null");
+                    }
+
+                    if (primaryKey is string stringKey)
+                    {
+                        return CreateObjectWithPrimaryKey(table, stringKey, update, out isNew);
+                    }
+
+                    throw new ArgumentException($"{parentType}'s primary key is defined as string, but the value passed is {primaryKey.GetType().Name}");
+                case PropertyType.Int:
+                    if (primaryKey == null)
+                    {
+                        if (!pkProperty.Type.IsNullable())
+                        {
+                            throw new ArgumentException($"{parentType}'s primary key is defined as non-nullable, but the value passed is null");
+                        }
+
+                        return CreateObjectWithPrimaryKey(table, update, out isNew);
+                    }
+
+                    var longKey = Convert.ToInt64(primaryKey);
+                    return CreateObjectWithPrimaryKey(table, longKey, pkProperty.Type.IsNullable(), update, out isNew);
+                default:
+                    throw new NotSupportedException($"Unexpected primary key of type: {pkProperty.Type}");
+            }
+        }
+
+        private IntPtr CreateObjectWithPrimaryKey(TableHandle table, long key, bool isNullable, bool update, out bool isNew)
+        {
+            var result = NativeMethods.create_object_unique(this, table, key, isNullable, update, out isNew, out var ex);
+            ex.ThrowIfNecessary();
+            return result;
+        }
+
+        private IntPtr CreateObjectWithPrimaryKey(TableHandle table, string key, bool update, out bool isNew)
+        {
+            var result = NativeMethods.create_object_unique(this, table, key, (IntPtr)key.Length, update, out isNew, out var ex);
+            ex.ThrowIfNecessary();
+            return result;
+        }
+
+        private IntPtr CreateObjectWithPrimaryKey(TableHandle table, bool update, out bool isNew)
+        {
+            var result = NativeMethods.create_object_unique(this, table, update, out isNew, out var ex);
+            ex.ThrowIfNecessary();
+            return result;
         }
 
         public class SchemaMarshaler